driver: enet: fix compatible issue for Cortex-A Core
Covert to be 64-bit compatible.

+++ drivers/enet/fsl_enet.c
@@ -2263,7 +2263,7 @@ status_t ENET_GetRxFrame(ENET_Type *base, enet_handle_t *handle, enet_rx_frame_s
         if (0U != (curBuffDescrip->control & ENET_BUFFDESCRIPTOR_RX_LAST_MASK))
         {
             /* The last buffer descriptor stores the error status of this received frame. */
-            result = ENET_GetRxFrameErr((enet_rx_bd_struct_t *)(uint32_t)curBuffDescrip, &rxFrame->rxFrameError);
+            result = ENET_GetRxFrameErr((enet_rx_bd_struct_t *)(uintptr_t)curBuffDescrip, &rxFrame->rxFrameError);
             break;
         }
